requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1
Pillow==10.2.0
typing-extensions==4.9.0
pydantic==2.6.1 